leetcode/0121.BestTimetoBuyandSellStock/121. Best Time to Buy and Sell Stock_test.go
package leetcode

import "testing"

func Test_maxProfit(t *testing.T) {
	tests := []struct {
		name   string
		prices []int
		want   int
	}{
		{
			name:   "example 1",
			prices: []int{7, 1, 5, 3, 6, 4},
			want:   5,
		},
		{
			name:   "example 2",
			prices: []int{7, 6, 4, 3, 1},
			want:   0,
		},
		{
			name:   "single price",
			prices: []int{1},
			want:   0,
		},
		{
			name:   "no prices",
			prices: []int{},
			want:   0,
		},
	}

	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := maxProfit(tt.prices); got != tt.want {
				t.Errorf("maxProfit() = %v, want %v", got, tt.want)
			}
			if got := maxProfit1(tt.prices); got != tt.want {
				t.Errorf("maxProfit() = %v, want %v", got, tt.want)
			}
			if got := maxProfit2(tt.prices); got != tt.want {
				t.Errorf("maxProfit() = %v, want %v", got, tt.want)
			}
			if got := maxProfit3(tt.prices); got != tt.want {
				t.Errorf("maxProfit() = %v, want %v", got, tt.want)
			}
		})
	}
}
